# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RetentionPolicyRequest": {
        "properties": {
          "retention_days": {
            "maximum": 3650.0,
            "minimum": 1.0,
            "title": "Retention Days",
            "type": "integer"
          }
        },
        "required": [
          "retention_days"
        ],
        "title": "RetentionPolicyRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Generated from the Agent Builder FastAPI application. Do not edit this file by hand; run scripts/generate_openapi.py.",
    "title": "Agent Builder Platform API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/brands/{brand_id}/privacy/retention": {
      "get": {
        "operationId": "get_retention_policy_api_v1_admin_brands__brand_id__privacy_retention_get",
        "parameters": [
          {
            "in": "path",
            "name": "brand_id",
            "required": true,
            "schema": {
              "title": "Brand Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-admin-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Retention Policy",
        "tags": [
          "admin",
          "privacy"
        ]
      },
      "put": {
        "operationId": "set_retention_policy_api_v1_admin_brands__brand_id__privacy_retention_put",
        "parameters": [
          {
            "in": "path",
            "name": "brand_id",
            "required": true,
            "schema": {
              "title": "Brand Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-admin-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetentionPolicyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Set Retention Policy",
        "tags": [
          "admin",
          "privacy"
        ]
      }
    },
    "/api/v1/admin/brands/{brand_id}/privacy/retention/run": {
      "post": {
        "operationId": "run_retention_cleanup_api_v1_admin_brands__brand_id__privacy_retention_run_post",
        "parameters": [
          {
            "in": "path",
            "name": "brand_id",
            "required": true,
            "schema": {
              "title": "Brand Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-admin-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Run Retention Cleanup",
        "tags": [
          "admin",
          "privacy"
        ]
      }
    },
    "/api/v1/admin/brands/{brand_id}/privacy/subjects/{subject_id}": {
      "delete": {
        "operationId": "delete_subject_data_api_v1_admin_brands__brand_id__privacy_subjects__subject_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "brand_id",
            "required": true,
            "schema": {
              "title": "Brand Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "subject_id",
            "required": true,
            "schema": {
              "title": "Subject Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-admin-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Subject Data",
        "tags": [
          "admin",
          "privacy"
        ]
      }
    },
    "/api/v1/admin/brands/{brand_id}/privacy/subjects/{subject_id}/export": {
      "get": {
        "operationId": "export_subject_data_api_v1_admin_brands__brand_id__privacy_subjects__subject_id__export_get",
        "parameters": [
          {
            "in": "path",
            "name": "brand_id",
            "required": true,
            "schema": {
              "title": "Brand Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "subject_id",
            "required": true,
            "schema": {
              "title": "Subject Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-admin-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Export Subject Data",
        "tags": [
          "admin",
          "privacy"
        ]
      }
    },
...
  }
}